[test/refactor] Refactored audit search tests
