+++ tests/integration/test_client.py
@@ -1,11 +1,13 @@
 import time
 from dataclasses import dataclass
 from typing import Generator, List, Optional, Type
+from unittest.mock import patch
 
 import pytest
 from pydantic.v1 import StrictStr
 
 from pyatlan.client.atlan import AtlanClient
+from pyatlan.client.audit import LOGGER
 from pyatlan.client.search_log import (
     AssetViews,
     SearchLogRequest,
@@ -22,7 +24,7 @@ from pyatlan.model.assets import (
     Database,
     Table,
 )
-from pyatlan.model.audit import AuditSearchRequest
+from pyatlan.model.audit import AuditSearchRequest, AuditSearchResults
 from pyatlan.model.core import Announcement
 from pyatlan.model.enums import (
     AnnouncementType,
@@ -599,7 +601,90 @@ def test_audit_find_by_user(
     assert audit_entity != audit_entity_next_page
 
 
+def _assert_audit_search_results(results, expected_sorts, size, total_count):
+    assert results.total_count > size
+    assert len(results.current_page()) == size
+    assert results.total_count == total_count
+    assert results
+    assert results._criteria.dsl.sort == expected_sorts
+
+
 @pytest.mark.order(after="test_audit_find_by_user")
+@patch.object(LOGGER, "debug")
+def test_audit_search_pagination(
+    mock_logger, audit_info: AuditInfo, client: AtlanClient
+):
+    size = 2
+
+    # Test audit search by GUID with default offset-based pagination
+    dsl = DSL(
+        query=Bool(filter=[Term(field="entityId", value=audit_info.guid)]),
+        sort=[],
+        size=size,
+    )
+    request = AuditSearchRequest(dsl=dsl)
+    results = client.audit.search(criteria=request, bulk=False)
+    total_count = results.total_count
+    expected_sorts = [SortItem(field="entityId", order=SortOrder.ASCENDING)]
+    _assert_audit_search_results(results, expected_sorts, size, total_count)
+
+    # Test audit search by guid with `bulk` option using timestamp-based pagination
+    results = client.audit.search(criteria=request, bulk=True)
+    total_count = results.total_count
+    expected_sorts = [
+        SortItem("created", order=SortOrder.ASCENDING),
+        SortItem(field="entityId", order=SortOrder.ASCENDING),
+    ]
+    _assert_audit_search_results(results, expected_sorts, size, total_count)
+    assert mock_logger.call_count == 1
+    assert "Audit bulk search option is enabled." in mock_logger.call_args_list[0][0][0]
+    mock_logger.reset_mock()
+
+    # When the number of results exceeds the predefined
+    # threshold and bulk is true and no pre-defined sort.
+    with patch.object(AuditSearchResults, "_MASS_EXTRACT_THRESHOLD", -1):
+        username = client.user.get_current().username
+        assert username
+        dsl = DSL(
+            query=Bool(filter=[Term(field="user", value=username)]),
+            sort=[],
+            size=size,
+        )
+        request = AuditSearchRequest(dsl=dsl)
+        results = client.audit.search(criteria=request, bulk=True)
+        total_count = results.total_count
+        expected_sorts = [
+            SortItem("created", order=SortOrder.ASCENDING),
+            SortItem(field="entityId", order=SortOrder.ASCENDING),
+        ]
+        _assert_audit_search_results(results, expected_sorts, size, total_count)
+        assert mock_logger.call_count < total_count
+        assert (
+            "Audit bulk search option is enabled."
+            in mock_logger.call_args_list[0][0][0]
+        )
+        mock_logger.reset_mock()
+
+    # When the number of results exceeds the predefined threshold and bulk is `False` and no pre-defined sort.
+    # Then SDK automatically switches to a `bulk` search option using timestamp-based pagination
+    with patch.object(AuditSearchResults, "_MASS_EXTRACT_THRESHOLD", -1):
+        request = AuditSearchRequest(dsl=dsl)
+        results = client.audit.search(criteria=request, bulk=False)
+        total_count = results.total_count
+        expected_sorts = [
+            SortItem("created", order=SortOrder.ASCENDING),
+            SortItem(field="entityId", order=SortOrder.ASCENDING),
+        ]
+        _assert_audit_search_results(results, expected_sorts, size, total_count)
+        assert mock_logger.call_count < total_count
+        assert (
+            "Result size (%s) exceeds threshold (%s)."
+            in mock_logger.call_args_list[0][0][0]
+        )
+        mock_logger.reset_mock()
+
+
+@pytest.mark.order(after="test_audit_search_pagination")
 def test_audit_find_by_qualified_name(client: AtlanClient, audit_info: AuditInfo):
     assert audit_info.qualified_name
     assert audit_info.type_name
