+++ tests/unit/test_audit_search.py
@@ -11,7 +11,6 @@ from pyatlan.model.enums import SortOrder
 from pyatlan.model.fluent_search import DSL
 from pyatlan.model.search import Bool, SortItem, Term
 
-# Constants
 SEARCH_RESPONSES_DIR = Path(__file__).parent / "data" / "search_responses"
 AUDIT_SEARCH_PAGING_JSON = "audit_search_paging.json"
 
@@ -34,18 +33,6 @@ def audit_search_paging_json():
             return load(input_file)
 
     return load_json(AUDIT_SEARCH_PAGING_JSON)
-
-
-TEST_THRESHOLD = 1
-
-
-def create_dsl(entity_id, sort=None, size=1, from_=0):
-    return DSL(
-        query=Bool(filter=[Term(field="entityId", value=entity_id)]),
-        sort=sort or [],
-        size=size,
-        from_=from_,
-    )
 
 
 def _assert_audit_search_results(results, response_json, sorts, bulk=False):
@@ -82,7 +69,7 @@ def test_audit_search_pagination(
         from_=0,
     )
     audit_search_request = AuditSearchRequest(dsl=dsl)
-    response = client.search(criteria=audit_search_request)
+    response = client.search(criteria=audit_search_request, bulk=False)
     expected_sorts = [SortItem(field="entityId", order=SortOrder.ASCENDING)]
 
     _assert_audit_search_results(response, audit_search_paging_json, expected_sorts)
@@ -90,12 +77,6 @@ def test_audit_search_pagination(
     mock_api_caller.reset_mock()
 
     # Test bulk pagination
-    dsl = DSL(
-        query=Bool(filter=[Term(field="entityId", value="some-guid")]),
-        sort=[],
-        size=2,
-        from_=0,
-    )
     audit_search_request = AuditSearchRequest(dsl=dsl)
     response = client.search(criteria=audit_search_request, bulk=True)
     expected_sorts = [
@@ -111,14 +92,13 @@ def test_audit_search_pagination(
     mock_logger.reset_mock()
     mock_api_caller.reset_mock()
 
-    # Test automatic bulk conversion when exceeding threshold
-    with patch.object(AuditSearchResults, "_MASS_EXTRACT_THRESHOLD", TEST_THRESHOLD):
+    # Test automatic bulk search conversion when exceeding threshold
+    with patch.object(AuditSearchResults, "_MASS_EXTRACT_THRESHOLD", -1):
         mock_api_caller._call_api.side_effect = [
             audit_search_paging_json,
             audit_search_paging_json,
             {},
         ]
-        dsl = create_dsl("some-guid")
         audit_search_request = AuditSearchRequest(dsl=dsl)
         response = client.search(criteria=audit_search_request)
         _assert_audit_search_results(
@@ -129,21 +109,34 @@ def test_audit_search_pagination(
             "Result size (%s) exceeds threshold (%s)"
             in mock_logger.call_args_list[0][0][0]
         )
+
+        # Test exception for bulk=False with user-defined sorting and results exceeds the predefined threshold
+        dsl.sort = dsl.sort + [SortItem(field="some-sort1", order=SortOrder.ASCENDING)]
+        audit_search_request = AuditSearchRequest(dsl=dsl)
+        with pytest.raises(
+            InvalidRequestError,
+            match=(
+                "ATLAN-PYTHON-400-066 Unable to execute "
+                "audit bulk search with user-defined sorting options. "
+                "Suggestion: Please ensure that no sorting options are "
+                "included in your search request when performing a bulk search."
+            ),
+        ):
+            client.search(criteria=audit_search_request, bulk=False)
+
+        # Test exception for bulk=True with user-defined sorting
+        dsl.sort = dsl.sort + [SortItem(field="some-sort2", order=SortOrder.ASCENDING)]
+        audit_search_request = AuditSearchRequest(dsl=dsl)
+        with pytest.raises(
+            InvalidRequestError,
+            match=(
+                "ATLAN-PYTHON-400-066 Unable to execute "
+                "audit bulk search with user-defined sorting options. "
+                "Suggestion: Please ensure that no sorting options are "
+                "included in your search request when performing a bulk search."
+            ),
+        ):
+            client.search(criteria=audit_search_request, bulk=True)
+
     mock_logger.reset_mock()
     mock_api_caller.reset_mock()
-
-    # Test exception for bulk=True with user-defined sorting
-    dsl = create_dsl(
-        "some-guid", sort=[SortItem(field="user", order=SortOrder.ASCENDING)]
-    )
-    audit_search_request = AuditSearchRequest(dsl=dsl)
-    with pytest.raises(
-        InvalidRequestError,
-        match=(
-            "ATLAN-PYTHON-400-066 Unable to execute "
-            "audit bulk search with user-defined sorting options. "
-            "Suggestion: Please ensure that no sorting options are "
-            "included in your search request when performing a bulk search."
-        ),
-    ):
-        client.search(criteria=audit_search_request, bulk=True)
